Add Parsing Objects
Primarily focused on adding the ability to parse a DatabaseValue & DatabaseResult to and from an object, instead of working with raw data. I don't know how far I want to go with advancing the library to "sync" data. This is a very, very manual process. Automating it would add ease of use, but also a high amount of complexity to the library.

* Added DatabaseResult#toObject
* Added DatabaseValue#fromObject
* Fixed major issue where TableSchema#toString would not include "TABLE" in "CREATE TABLE" if ifNotExists and orReplace were both set to false.